+++ Nursery Management System WPF/Nursery Management System WPF/Classes/SQLQuery.cs
@@ -529,7 +529,7 @@ namespace Nursery_Management_System_WPF
 
                 currentRoom.id = Convert.ToInt32(dr["roomID"].ToString());
                 currentRoom.number = Convert.ToInt32(dr["roomNumber"].ToString());
-                currentRoom.staffID = Convert.ToInt32(dr["roomStaffID"].ToString());
+                currentRoom.staffID = Convert.ToInt64(dr["roomStaffID"].ToString());
 
                 room.AddLast(currentRoom);
             }
